+++ backend/agent/agent.py
@@ -1,8 +1,11 @@
 # import os
+import keyword
 import openai
 import json
+import os
 from typing import Any, List, Optional, Callable
 from pydantic import BaseModel
+from database.my_codebase import MyCodebase
 
 # from agent.agent_functions import Program, File
 
@@ -28,23 +31,55 @@ class Message(BaseModel):
 
 
 class CodingAgent:
+    """
+    A class to represent a coding agent that uses OpenAI's GPT-3 model to generate code.
+
+    Attributes:
+    memory_manager : Manages the memory of the agent.
+    functions : A list of functions that the agent can call.
+    callables : A list of callable functions.
+    GPT_MODEL : The GPT-3 model used by the agent.
+    function_map : A dictionary mapping function names to their callable objects.
+    """
+
     def __init__(
         self,
         memory_manager,
         functions: Optional[List[dict]] = None,
         callables: Optional[List[Callable]] = None,
+        codebase: Optional[MyCodebase] = None,
     ):
+        """
+        Constructs all the necessary attributes for the CodingAgent object.
+
+        Parameters:
+        memory_manager : Manages the memory of the agent.
+        functions (Optional[List[dict]]) : A list of functions that the agent can call.
+        callables (Optional[List[Callable]]) : A list of callable functions.
+        """
         self.memory_manager = memory_manager
         self.functions = functions
         self.callables = callables
         self.GPT_MODEL = GPT_MODEL
+        self.codebase = codebase
         if callables:
             self.function_map = {
                 func.__name__: func for func in callables if func is not None
             }
+        self.files_in_prompt: List[str] = []
         print(f"Function Map: {self.function_map}")
 
     def query(self, input: str, command: Optional[str] = None) -> List[str]:
+        """
+        Queries the GPT-3 model with the given input and command.
+
+        Parameters:
+        input (str) : The input text to be processed by the GPT-3 model.
+        command (Optional[str]) : The command to be executed by the agent.
+
+        Returns:
+        List[str] : The output generated by the GPT-3 model.
+        """
         print(f"Input Text: {input}")
         self.memory_manager.add_message("user", input)
         message_history = [
@@ -71,6 +106,14 @@ class CodingAgent:
             keyword_args["functions"] = [self.function_map.get(command).openai_schema]
             keyword_args["function_call"] = {"name": command}
 
+            if command == "Changes":
+                self.memory_manager.identity = (
+                    self.memory_manager.identity
+                    + "\nLine numbers have been added to the Current File to aid in your response. They are not part of the actual file."
+                )
+                self.set_files_in_prompt(include_line_numbers=True)
+                keyword_args["model"] = "gpt-4"
+
         for chunk in openai.ChatCompletion.create(**keyword_args):
             delta = chunk["choices"][0].get("delta", {})
             if "function_call" in delta:
@@ -86,8 +129,8 @@ class CodingAgent:
                 args = json.loads(function_to_call.arguments)
                 function_response = self.function_map[function_to_call.name](**args)
                 print(f"Func Response: {json.dumps(function_response.to_dict())}")
-                if function_to_call.name == "FileChange":
-                    diff = function_response.save()
+                if function_to_call.name == "Changes":
+                    diff = function_response.execute()
                     # Show the diff back to the user
                     yield diff
 
@@ -109,3 +152,42 @@ class CodingAgent:
                     yield content
             else:
                 yield delta.get("content")
+
+    def set_files_in_prompt(self, include_line_numbers: Optional[bool] = None) -> None:
+        """
+        Sets the files in the prompt.
+
+        Parameters:
+        files (List[File]) : A list of files to be set in the prompt.
+        """
+        file_contents = self.codebase.get_file_contents()
+        files = [
+            os.path.join(self.codebase.directory, f.lstrip("/"))
+            for f in self.files_in_prompt
+        ]
+        content = ""
+        for k, v in file_contents.items():
+            if k in files and include_line_numbers:
+                v = self.add_line_numbers_to_content(v)
+                content += f"{k}:\n{v}\n\n"
+            elif k in files:
+                content += f"{k}:\n{v}\n\n"
+
+        self.memory_manager.system_file_contents = content
+        self.memory_manager.set_system()
+        return
+
+    def add_line_numbers_to_content(self, content: str) -> str:
+        """
+        Adds line numbers to the given content.
+
+        Parameters:
+        content (str) : The content to add line numbers to.
+
+        Returns:
+        str : The content with line numbers added.
+        """
+        lines = content.split("\n")
+        for i in range(len(lines)):
+            lines[i] = f"{i+1} {lines[i]}"
+        return "\n".join(lines)
