+++ backend/tests/test_agent_functions.py
@@ -1,72 +1,61 @@
 import os
-import tempfile
-from agent.agent_functions import FileChange
-from diff_match_patch import diff_match_patch
+import shutil
+from app_setup import setup_app_testing, DIRECTORY
 
-
-def test_FileChange():
-    # Create a temporary file
-    with tempfile.NamedTemporaryFile(mode="w+", delete=False) as temp_file:
-        temp_file.write("Hello\n")
-
-    # Create a FileChange instance
-    file_change = FileChange(name=temp_file.name, changes="Change 'Hello' to 'Hi'")
-
-    # # Save the changes
-    file_change.save()
-
-    # # Check that the file was updated correctly
-    with open(temp_file.name, "r") as f:
-        assert f.read() == "Hi\n"
-
-    # Clean up the temporary file
-    os.remove(temp_file.name)
+AGENT, CODEBASE = setup_app_testing()
+temp_file = "backend/tests/test_files/agent_function_test1"
+temp_file_full = os.path.join(DIRECTORY, temp_file)
 
 
 def test_FileChange_multiple_changes():
     # Create a temporary file
-    with tempfile.NamedTemporaryFile(mode="w+", delete=False) as temp_file:
-        temp_file.write("Hello\nWorld\n")
+    with open(temp_file_full, "w") as f:
+        f.write("Hello\nWorld\n")
 
     # Create a FileChange instance
-    file_change = FileChange(
-        name=temp_file.name, changes="Change 'Hello' to 'Hi' and 'World' to 'Everyone'"
-    )
-
-    # Save the changes
-    new_text = file_change.save()
+    accumulated_message = ""
+    for content in AGENT.query(
+        input=f"Change 'Hello' to 'Hi' and 'World' to 'Everyone' in {temp_file}",
+        command="Changes",
+    ):
+        if content is not None:
+            accumulated_message += content
 
     # Check that the file was updated correctly
-    with open(temp_file.name, "r") as f:
+    with open(temp_file_full, "r") as f:
         assert f.read() == "Hi\nEveryone\n"
 
     # Clean up the temporary file
-    os.remove(temp_file.name)
+    os.remove(temp_file_full)
 
 
 def test_FileChange_real_world_example():
     # Create a temporary file
-    TEST_FILE = "tests/test_files/app_setup_test.py"
-
-    with tempfile.NamedTemporaryFile(mode="w+", delete=False) as temp_file:
-        temp_file.write(open(TEST_FILE).read())
+    TEST_FILE = "backend/tests/test_files/app_setup_test.py"
+    FULL_PATH = os.path.join(DIRECTORY, TEST_FILE)
+    shutil.copy2(FULL_PATH, temp_file_full)
+    AGENT.files_in_prompt = [TEST_FILE]
+    AGENT.set_files_in_prompt(include_line_numbers=True)
 
     # Create a FileChange instance
-    file_change = FileChange(
-        name=temp_file.name,
-        changes="Change 'Successfully connected to database' to 'Database connection established', add a comment '# This is a new comment' at the end of the 'create_database_connection' function, and delete the duplicate 'DB_CONNECTION = create_database_connection()' line",
-    )
+    accumulated_message = ""
+    for content in AGENT.query(
+        input=f"""
+        For file - {temp_file}, Change 'Successfully connected to database' to 'Database connection established',
+        add a comment '# This is a new comment' after the 'create_database_connection' function, and finally
+        delete the duplicate 'DB_CONNECTION = create_database_connection()' line",
+        command="Changes""",
+    ):
+        if content is not None:
+            accumulated_message += content
 
-    # Save the changes
-    _ = file_change.save()
     # Check that the file was updated correctly
-    with open(temp_file.name, "r") as f:
+    with open(temp_file_full, "r") as f:
         new_text = f.readlines()
-        new_text
-
-    assert new_text[31].strip() == 'print("Database connection established")'
 
     # Should be 8 spaces for indents and one newline character
-    assert len(new_text[31]) - len('print("Database connection established")') == 9
-    assert new_text[45].strip() == "# This is a new comment"
-    assert new_text[68].strip() == ""
+    assert len(new_text[32]) - len('print("Database connection established")') == 9
+    assert new_text[32].strip() == 'print("Database connection established")'
+    assert new_text[46].strip() == "# This is a new comment"
+    assert new_text[69].strip() == ""
+    os.remove(temp_file_full)
